Clean-up duplicated PathInternal files (#57988)
* Clean-up duplicated PathInternal files

* Move PathInternal*.cs to Common
