+++ src/libraries/Common/src/System/IO/PathInternal.cs
